// src/api/apiTypes.ts
/**
 * file is autogenerated by utils/generateApi.ts, do not edit
 */

/* eslint-disable camelcase, no-use-before-define, @typescript-eslint/no-unused-vars */

export interface Error {
  message: string
  code?: string | null
  fields?:
    | {
        [key: string]: unknown
      }[]
    | null
}

export type AuthUserInfo = AuthToken & UserInfo

export interface AuthToken {
  token: string
}

export interface UserInfo {
  id: number
  email: string
  firstName: string
  lastName: string
  phone?: string | null
  image?: string | null
}

export interface AuthCredentials {
  username: string
  password: string
}

export type UsersPagedList = PagedList & UserList

export interface PagedList {
  limit: number
  skip: number
  total: number
}

export interface UserList {
  users: UserInfo[]
}

export interface UserForm {
  email: string
  firstName: string
  lastName: string
  phone?: string | null
  image?: string | null
}